Splits functional tests into different files
