+++ .gitignore
@@ -15,3 +15,4 @@ test.conf
 .venv/
 .direnv/
 .envrc
+*.lic
