+++ tests/sample.conf
@@ -6,3 +6,4 @@ ecs_endpoint = https://10.0.1.1:4443
 username = root
 password = ChangeMe
 api_version = 3
+license_file = /home/user/license.lic
